ref: Goal
Redefined goal of the repo. The aim is to focus on the repo being solely a template without the walkthrough element. Instaead, the walkthrough element will be replaced with a concise summary explaining what has been set up and why.

The goal has been redefined to make the repo objective clearer.